requirements.txt
python-dotenv
celery==4.4.5
pika==1.0.0
slackclient
